+++ ui/readyhandlers.js
@@ -93,6 +93,7 @@ $(document).ready(function(){
     if (prefs.sync_lastfm_at_start) {
         syncLastFMPlaycounts.start();
     }
+    spotifyLinkChecker.initialise();
 });
 
 function cleanBackendCache() {